Set up a types module for search-index-new feat. module
- all types also exported from feat. module index
- moved some out-of-place stuff into there
