+++ src/search/search-index-new/index.ts
@@ -21,6 +21,7 @@ const index = new Proxy<Storage>({} as Storage, {
 
 export const init = (props?: Props) => (realIndex = new Storage(props))
 
+export * from './types'
 export { Storage }
 export default index
